% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enzyme_assayClass.R
\name{enzyme_assay.calc_slope}
\alias{enzyme_assay.calc_slope}
\title{Getting Slope from Range}
\usage{
enzyme_assay.calc_slope(object)
}
\arguments{
\item{object}{An object of the Enzyme_Assay class}
}
\description{
This internal function calculates the slopes associated with curated ranges
}
\examples{
ea <- enzyme_assay(enzyme_assay.data, enzyme_assay.time)
eaCurated <- enzyme_assay.curate(ea)
}

