% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enzyme_assayClass.R
\name{enzyme_assay.curate}
\alias{enzyme_assay.curate}
\title{Well-By-Well Curation}
\usage{
enzyme_assay.curate(x, ...)
}
\arguments{
\item{x}{An object of the Enzyme_Assay class}

\item{at_once}{Number of wells/columns to plot at once}

\item{status}{Whether to start from 'auto' or 'curated' ranges}
}
\description{
This function plots and asks the user to curate an Enzyme_Assay object well by well
}
\examples{
ea <- enzyme_assay(enzyme_assay.data, enzyme_assay.time)
eaCurated <- enzyme_assay.curate(ea)
}

