% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enzyme_assayClass.R
\name{enzyme_assay.data}
\alias{enzyme_assay.data}
\title{Example Enzyme Assay Output}
\format{A data frame with 69 rows and 42 columns}
\description{
An example enzyme assay run with absorbance values.
Rows for different measurement time points and columns
for different wells.
}

