% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enzyme_assayClass.R
\name{enzyme_assay.s_to_measurement}
\alias{enzyme_assay.s_to_measurement}
\title{Finding Measurement Closest to Seconds}
\usage{
enzyme_assay.s_to_measurement(this, s)
}
\arguments{
\item{this}{An object of the Enzyme_Assay class}

\item{s}{A number in seconds}
}
\description{
This function returns the measurement closest to a selected seconds value
}
\examples{
ea <- enzyme_assay(enzyme_assay.data, enzyme_assay.time)
enzyme_assay.s_to_measurement(ea, 111)
}

