% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/enzyme_assayClass.R
\name{enzyme_assay.time}
\alias{enzyme_assay.time}
\title{Example Plate Reader Timepoints}
\format{A data frame with 69 rows and 42 columns}
\description{
An example vector of time values produced by the plate
reader that corresponds the the measurements in
enzyme_assay.time
}

