% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/plate_reader_seconds.R
\name{plate_reader_seconds}
\alias{plate_reader_seconds}
\title{Convert to Seconds}
\usage{
plate_reader_seconds(time)
}
\arguments{
\item{time}{a value or vector of strings for time points in the format "H:MM:SS"}
}
\description{
This function converts time in the format exported by the plate reader (H:MM:SS) to seconds
}
\examples{
plate_reader_seconds(enzyme_assay.time)
}

