% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/wrap_col.R
\name{wrap_col}
\alias{wrap_col}
\title{Replicate Vector to N}
\usage{
wrap_col(col, n)
}
\arguments{
\item{col}{A vector of colors.}

\item{n}{The desired length for the returned vector of colors.}
}
\description{
This function extends a vector of colors to specified length.
}
\examples{
wrap_col(c('red','blue'), 8)
}

