% Generated by roxygen2 (4.1.0.9001): editted by hand anyways
% Please edit documentation in R/enzyme_assayClass.R 
\docType{class}
\name{Enzyme_Assay}
\alias{Enzyme_Assay}
\title{Organizing Enzyme Assay Data}
\description{
This class holds, and facilitates the curation of enzyme assay data
}
\section{Slots}{

\describe{
\item{\code{data}}{A matrix with measurements in rows for different wells in columns}

\item{\code{seconds}}{The second values corresponding to the rows in slot 'data'}

\item{\code{auto_slope}}{The automatically calculated maximum slopes from \code{\link{enzyme_assay}}}

\item{\code{auto_range}}{The ranges corresponding to slopes in slot 'auto_slope'}

\item{\code{curated_slope}}{The user curated slopes from \code{\link{enzyme_assay.curate}} or \code{\link{enzyme_assay.curate_batch}}}

\item{\code{curated_range}}{The ranges corresponding to slot 'curated_slope'}
}}
\examples{
ea <- enzyme_assay(enzyme_assay.data, enzyme_assay.time)
eaCurated <- enzyme_assay.curate(ea)
new_range <- ea@auto_range
new_range[1,] <- 1
eaBatchCurated <- enzyme_assay.curate_batch(ea, new_range)
}

